pyproject.toml
[tool.poetry]
name = "template-flask"
version = "0.1.0"
description = ""
authors = ["Marcelo Lino <mdslino@gmail.com>"]
readme = "README.md"
packages = []

[tool.poetry.dependencies]
Flask = "^2.2.2"
Flask-Admin = "^1.6.0"
Flask-RESTful = "^0.3.9"
Flask-SQLAlchemy = "^2.5.1"
pydantic = { extras = ["email"], version = "^1.10.2" }
python = "^3.10"
gunicorn = "^20.1.0"
gevent = "^21.12.0"
Bootstrap-Flask = "^2.1.0"
psycopg2-binary = "^2.9.3"
Flask-Migrate = "^3.1.0"
Flask-Login = "^0.6.2"
Flask-WTF = "^1.0.1"
sqlalchemy-utils = "^0.39.0"


[tool.poetry.group.dev.dependencies]
black = "^22.8.0"
flake8 = "^5.0.4"
Flask-DebugToolbar = "^0.13.1"
flask-shell-ipython = "^0.4.1"
ipdb = "^0.13.9"
isort = "^5.10.1"
mypy = "^0.971"
python-dotenv = "^0.21.0"


[tool.poetry.group.test.dependencies]
pytest-flask = "^1.2.0"
pytest-cov = "^4.0.0"
factory-boy = "^3.2.1"
pytest-dotenv = "^0.5.2"

[tool.isort]
profile = "black"
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true
line_length = 79

[tool.black]
line-length = 79
target-version = ['py310']
include = '\.pyi?$'
extend-exclude = '''
/(
  # The following are specific to Black, you probably don't want those.
  | blib2to3
  | tests/data
  | profiling
)/
'''
# We use preview style for formatting Black itself. If you
# want stable formatting across releases, you should keep
# this off.
preview = true

[tool.pytest.ini_options]
env_files = [
    ".test.env"
]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
